requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
eventlet==0.34.2
ansible==8.7.0
PyYAML==6.0.1
APScheduler==3.10.4
SQLAlchemy==2.0.23
pymongo==4.6.1
requests==2.31.0
